// generated/src/app/api-models/trading-partner-detail.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class TradingPartnerDetail {
  tradingPartnerId: string;
  transactionType: string;
  description: string;
  autoValidate: string;
  autoPost: string;
  partialPost: string;
  reworkable: string;
  processByBatch: string;
  priceClaims: string;
  adjudicateClaims: string;
  checkForDuplicates: string;
  paymentForIndividual: string;
  userDefined1: string;
  userDefined2: string;
  comments: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}